requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.0.5
Flask-Login==0.6.2
Flask-CORS==4.0.0
python-dotenv==1.0.0
geopy==2.4.0
bcrypt==4.0.1
Werkzeug==2.3.7  # Specify compatible version